+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Angular Gamification Changelog
 
+<a name="Fev 17, 2018"></a>
+### Fev 17, 2018
+* Mission reports
+
 <a name="Fev 16, 2018"></a>
 ### Fev 16, 2018
 * Accepting module config